refactor(multigateway): address PR review feedback
Move ExtractSQLSTATE and ClassifyErrorSource to mterrors package for
reusability. Inline ExtractOperationName at call sites. Move
defaultSlowQueryThreshold to constants package. Remove Vitess references
from documentation.
